requirements.txt
llama-cpp-python==0.3.7
sqlite-vec==0.1.1
streamlit==1.38.0